Update GenreRestController
add 'negative ID check' in 'deleteById' method;
add corresponding test;

+++ src/main/java/org/maximsavin/api_bookstore/domain/genre/GenreRestController.java
@@ -54,6 +54,8 @@ public class GenreRestController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Object> deleteById(@PathVariable long id) {
+        if (id < 0)
+            return ResponseEntity.badRequest().body("ID cannot be negative.");
         try {
             service.deleteById(id);
             return ResponseEntity.noContent().build();
